requirements.txt
Flask==3.0.2
flask_cors==4.0.0
langchain==0.1.9
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.29.0
streamlit_chat==0.1.1
